+++ FlexLib/Classes/FlexScrollView.m
@@ -11,24 +11,14 @@
 
 #import "FlexScrollView.h"
 #import "FlexRootView.h"
-#import "FlexParentView.h"
 #import "ViewExt/UIView+Flex.h"
 #import "YogaKit/UIView+Yoga.h"
 #import "FlexUtils.h"
 #import "FlexTouchView.h"
 
-static void* gObserverFrame = &gObserverFrame;
-static void* gObserverContentOffset = &gObserverContentOffset;
-
 @interface FlexScrollView()
 {
-    FlexParentView* _subview;
     FlexRootView* _contentView;
-    UIView* _holder;
-    
-    UIView* _stickingView;
-    NSInteger _stickingZIndex;
-    BOOL _stickingYChanged;
 }
 @end
 
@@ -39,20 +29,8 @@ static void* gObserverContentOffset = &gObserverContentOffset;
 {
     self = [super init];
     if (self) {
-        
-        _stickingZIndex = -1;
-        
-        // 占位的view
-        _holder = [[UIView alloc]init];
-        [_holder enableFlexLayout:YES];
-        [super addSubview:_holder];
-        
         // 构造view tree
         __weak FlexScrollView* weakSelf = self;
-        _subview = [[FlexParentView alloc]init];
-        _subview.onFrameChange = ^(CGRect rc){
-            weakSelf.contentSize = rc.size ;
-        };
         
         _contentView = [[FlexRootView alloc]init];
         _contentView.onWillLayout = ^{
@@ -61,19 +39,13 @@ static void* gObserverContentOffset = &gObserverContentOffset;
         _contentView.onDidLayout = ^{
             [weakSelf onContentViewDidLayout];
         };
-        [_subview addSubview:_contentView];
-        [super addSubview:_subview];
-        
-        [self addObserver:self forKeyPath:@"frame" options:NSKeyValueObservingOptionNew|NSKeyValueObservingOptionOld context:gObserverFrame];
-        [self addObserver:self forKeyPath:@"contentOffset" options:NSKeyValueObservingOptionNew context:gObserverContentOffset];
+        [super addSubview:_contentView];
     }
     return self;
 }
 
 - (void)dealloc
 {
-    [self removeObserver:self forKeyPath:@"frame"];
-    [self removeObserver:self forKeyPath:@"contentOffset"];
 }
 -(FlexRootView*)contentView
 {
@@ -81,41 +53,32 @@ static void* gObserverContentOffset = &gObserverContentOffset;
 }
 -(void)onContentViewWillLayout
 {
-    [self restoreStickView];
+
 }
 
 -(void)onContentViewDidLayout
 {
-    if(_stickingView != nil){
-        _stickingView.viewAttrs.originY = _stickingView.frame.origin.y ;
-        [self resetStickViews:YES];
+    if (!CGSizeEqualToSize(self.contentSize,_contentView.frame.size))
+    {
+        self.contentSize = _contentView.frame.size;
     }
 }
 
-#pragma mark - KVO
-
-- (void)observeValueForKeyPath:(NSString *)keyPath ofObject:(UIView*)object change:(NSDictionary *)change context:(void *)context
+- (void)setFrame:(CGRect)frame
 {
-    if(context == gObserverFrame){
-        
-        CGSize szNew = [[change objectForKey:@"new"]CGRectValue].size;
-        CGSize szOld = [[change objectForKey:@"old"]CGRectValue].size;
-        
-        if(!CGSizeEqualToSize(szOld, szNew)){
-            CGRect rc = _subview.frame;
-            if(!self.vertical)
-                rc.size.height = szNew.height;
-            if(!self.horizontal)
-                rc.size.width = szNew.width;
-            if(!CGSizeEqualToSize(rc.size,_subview.frame.size)){
-                _subview.frame = rc;
-                [_contentView setNeedsLayout];
-            }
+    CGSize oldSize = self.frame.size;
+    
+    [super setFrame:frame];
+    
+    if(!CGSizeEqualToSize(oldSize, frame.size)){
+        CGRect rc = _contentView.frame;
+        if(!self.vertical)
+            rc.size.height = CGRectGetHeight(frame);
+        if(!self.horizontal)
+            rc.size.width = CGRectGetWidth(frame);
+        if(!CGSizeEqualToSize(rc.size,_contentView.frame.size)){
+            _contentView.frame = rc;
         }
-    }else if(context == gObserverContentOffset){
-        [self resetStickViews:NO];
-    }else{
-        [super observeValueForKeyPath:keyPath ofObject:object change:change context:context];
     }
 }
 
@@ -161,10 +124,6 @@ if(from.prop.unit==YGUnitPoint||    \
 }
 -(void)addSubview:(UIView *)view
 {
-    if(view.viewAttrs.stickTop){
-        _stickingZIndex = _contentView.subviews.count;
-        _stickingView = view;
-    }
     [_contentView addSubview:view];
     [_contentView registSubView:view];
 }
@@ -173,37 +132,7 @@ if(from.prop.unit==YGUnitPoint||    \
     [view removeFromSuperview];
     [_contentView removeWatchView:view];
 }
--(void)restoreStickView
-{
-    if(_stickingView == nil||!_stickingYChanged)
-        return;
-    
-    CGRect rc = _stickingView.frame ;
-    rc.origin.y = _stickingView.viewAttrs.originY;
-    _stickingView.frame = rc;
-    [_stickingView.yoga markDirty];
-    [_contentView insertSubview:_stickingView atIndex:_stickingZIndex];
-    _stickingYChanged = NO;
-}
--(void)resetStickViews:(BOOL)fromLayout
-{
-    if(_stickingView == nil)
-        return;
-    
-    CGFloat offsetY = self.contentOffset.y ;
-    CGFloat y = _stickingView.viewAttrs.originY ;
-    CGFloat newY = MAX(offsetY,y);
-    
-    CGRect rc = _stickingView.frame ;
-    
-    if(rc.origin.y != newY){
-        rc.origin.y = newY ;
-        _stickingView.frame = rc;
-        
-        [_contentView bringSubviewToFront:_stickingView];
-        _stickingYChanged = YES;
-    }
-}
+
 FLEXSET(horzScroll)
 {
     BOOL b = String2BOOL(sValue) ;