+++ macosx/Classes/FlexRootView.m
@@ -237,6 +237,7 @@ static NSInteger _compareInputView(NSView * _Nonnull f,
 }
 @end
 
+
 @interface FlexRootView()
 {
     BOOL _bInLayouting;
@@ -260,8 +261,6 @@ static NSInteger _compareInputView(NSView * _Nonnull f,
         _safeArea = NSEdgeInsetsMake(0, 0, 0, 0);
         _lastConfigFrame = CGRectZero;
         _bChildDirty = NO;
-        
-        [self enableFlexLayout:YES];
     }
     return self;
 }
@@ -277,11 +276,15 @@ static NSInteger _compareInputView(NSView * _Nonnull f,
     return _owner;
 }
 
-+(FlexRootView*)loadWithNodeData:(NSData*)data
-                           Owner:(NSObject*)owner
+- (BOOL)inLayouting
 {
-    FlexRootView* root = [[FlexRootView alloc]init];
-    root->_owner = owner;
+    return _bInLayouting;
+}
+
+-(BOOL)loadWithNodeData:(NSData*)data
+                  Owner:(NSObject*)owner
+{
+    _owner = owner;
     
     FlexNode* node = [FlexNode loadNodeData:data];
     if(node != nil){
@@ -289,24 +292,25 @@ static NSInteger _compareInputView(NSView * _Nonnull f,
         
         @try{
             sub = [node buildViewTree:owner
-                             RootView:root];
+                             RootView:self];
         }@catch(NSException* exception){
             NSLog(@"Flexbox: FlexRootView exception occured - %@",exception);
         }
         
-        [root addSubview:sub];
+        [self addSubview:sub];
+        return sub!=nil;
     }
-    return root;
+    return NO;
 }
-+(FlexRootView*)loadWithNodeFile:(NSString*)resName
-                           Owner:(NSObject*)owner
+
+-(BOOL)loadWithNodeFile:(NSString*)resName
+                  Owner:(NSObject*)owner
 {
     if(resName==nil){
         resName = NSStringFromClass([owner class]);
     }
     
-    FlexRootView* root = [[FlexRootView alloc]init];
-    root->_owner = owner;
+    _owner = owner;
     
     FlexNode* node = [FlexNode loadNodeFromRes:resName Owner:owner];
     if(node != nil){
@@ -314,13 +318,29 @@ static NSInteger _compareInputView(NSView * _Nonnull f,
         
         @try{
             sub = [node buildViewTree:owner
-                             RootView:root];
+                             RootView:self];
         }@catch(NSException* exception){
             NSLog(@"Flexbox: FlexRootView exception occured - %@",exception);
         }
         
-        [root addSubview:sub];
+        [self addSubview:sub];
+        return sub!=nil;
     }
+    return NO;
+}
+
++(FlexRootView*)loadWithNodeData:(NSData*)data
+                           Owner:(NSObject*)owner
+{
+    FlexRootView* root = [[FlexRootView alloc]init];
+    [root loadWithNodeData:data Owner:owner];
+    return root;
+}
++(FlexRootView*)loadWithNodeFile:(NSString*)resName
+                           Owner:(NSObject*)owner
+{
+    FlexRootView* root = [[FlexRootView alloc]init];
+    [root loadWithNodeFile:resName Owner:owner];
     return root;
 }
 
@@ -429,20 +449,27 @@ static NSInteger _compareInputView(NSView * _Nonnull f,
 
 -(CGRect)getSafeArea
 {
-    CGRect rcSafeArea = self.superview.frame ;
-    rcSafeArea.origin = CGPointZero;
+    CGRect rcSafeArea ;
     
-    NSEdgeInsets inset;
-    if(self.calcSafeArea!=nil){
-        inset = self.calcSafeArea();
-    }else{
-        inset = self.safeArea;
-    }
-    
-    return CGRectMake(rcSafeArea.origin.x+inset.left,
+    if (self.useFrame) {
+        rcSafeArea = self.frame;
+    } else {
+        rcSafeArea = self.superview.frame ;
+        rcSafeArea.origin = CGPointZero;
+        
+        NSEdgeInsets inset;
+        if(self.calcSafeArea!=nil)
+        {
+            inset = self.calcSafeArea();
+        }else{
+            inset = self.safeArea;
+        }
+        rcSafeArea = CGRectMake(rcSafeArea.origin.x+inset.left,
                       rcSafeArea.origin.y+inset.top,
                       rcSafeArea.size.width-inset.left-inset.right,
                       rcSafeArea.size.height-inset.top-inset.bottom);
+    }
+    return rcSafeArea;
 }
 -(BOOL)isConfigSame
 {
@@ -483,7 +510,9 @@ static NSInteger _compareInputView(NSView * _Nonnull f,
         self.onWillLayout();
     }
     
+    [self enableFlexLayout:YES];
     [self.yoga applyLayoutPreservingOrigin:NO dimensionFlexibility:option];
+    [self enableFlexLayout:NO];
     
     // 布局后事件
     if(self.onDidLayout != nil){
@@ -518,7 +547,10 @@ static NSInteger _compareInputView(NSView * _Nonnull f,
     if(self.flexibleHeight)
         szLimit.height = NAN ;
     
-    CGSize sz=[self.yoga calculateLayoutWithSize:szLimit];
+    CGSize sz;
+    [self enableFlexLayout:YES];
+    sz=[self.yoga calculateLayoutWithSize:szLimit];
+    [self enableFlexLayout:NO];
     return sz;
 }
 -(CGSize)calculateSize
@@ -534,7 +566,7 @@ static NSInteger _compareInputView(NSView * _Nonnull f,
 //        [NSView beginAnimations:nil context:nil];
 //        [NSView setAnimationDuration:duration];
 //    };
-//    
+//
 //    self.endLayout = ^{
 //        [NSView commitAnimations];
 //    };